Added casts to eliminate warnings from WWPVRTCImage.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@1141 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWindIOS/WorldWind/WorldWind/Formats/PVRTC/WWPVRTCImage.m
@@ -928,7 +928,9 @@ unsigned long PackPvrBlock(RawImage* imageLo, RawImage* imageHi, RawImage* image
         }
     }
 
-    unsigned long block = ((unsigned long) wts) | (((unsigned long) rgbHi) << 48) | (((unsigned long) rgbLo) << 32);
+    unsigned long long block = ((unsigned long) wts) | (((unsigned long long) rgbHi) << 48) | (((unsigned long long)
+    rgbLo) <<
+        32);
     return block;
 }
